+++ tests/cadence/_internal/workflow/test_context_retry_policy.py
@@ -5,6 +5,7 @@ import pytest
 from google.protobuf.duration_pb2 import Duration
 
 from cadence._internal.workflow.context import Context
+from cadence._internal.workflow.deterministic_event_loop import DeterministicEventLoop
 from cadence.api.v1.common_pb2 import ActivityType
 from cadence.api.v1.decision_pb2 import ScheduleActivityTaskDecisionAttributes
 from cadence.api.v1.tasklist_pb2 import TaskList, TaskListKind
@@ -27,7 +28,7 @@ async def test_execute_activity_passes_retry_policy_to_schedule():
         workflow_task_list="tl",
         data_converter=DefaultDataConverter(),
     )
-    ctx = Context(info, dm)
+    ctx = Context(info, dm, DeterministicEventLoop())
 
     await ctx.execute_activity(
         "Act",
